# src/arxiv_cortex/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arxiv Cortex read-only API",
    "version": "1.0.0",
    "description": "Local paper discovery API. All endpoints are read-only and return UTC ISO-8601 timestamps."
  },
  "servers": [{"url": "/api/v1"}],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Inspect local index health",
        "responses": {"200": {"description": "Health and index counts"}}
      }
    },
    "/papers": {
      "get": {
        "operationId": "listPapers",
        "summary": "Search and list papers",
        "parameters": [
          {"name": "q", "in": "query", "description": "Plain words are combined with AND. Comma-, semicolon-, or newline-separated groups are matched as exact phrases and combined with OR.", "schema": {"type": "string"}},
          {"name": "category", "in": "query", "schema": {"type": "string"}},
          {"name": "days", "in": "query", "schema": {"type": "integer", "minimum": 0}},
          {"name": "state", "in": "query", "schema": {"type": "string", "enum": ["saved", "read", "unread", "dismissed"]}},
          {"name": "sort", "in": "query", "schema": {"type": "string", "enum": ["newest", "oldest", "relevance", "recommended"]}},
          {"name": "cursor", "in": "query", "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 25}}
        ],
        "responses": {"200": {"description": "A page of papers"}, "400": {"$ref": "#/components/responses/BadRequest"}}
      }
    },
    "/papers/{arxiv_id}": {
      "get": {
        "operationId": "getPaper",
        "summary": "Get one paper",
        "parameters": [{"$ref": "#/components/parameters/ArxivId"}],
        "responses": {"200": {"description": "Paper metadata"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/papers/{arxiv_id}/similar": {
      "get": {
        "operationId": "findSimilarPapers",
        "summary": "Rank papers by exact cosine similarity",
        "parameters": [
          {"$ref": "#/components/parameters/ArxivId"},
          {"name": "category", "in": "query", "schema": {"type": "string"}},
          {"name": "days", "in": "query", "schema": {"type": "integer", "minimum": 0}},
          {"name": "cursor", "in": "query", "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 25}}
        ],
        "responses": {"200": {"description": "Similarity-ranked papers"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/recommendations": {
      "get": {
        "operationId": "getRecommendations",
        "summary": "Get personalized recommendations",
        "parameters": [
          {"name": "category", "in": "query", "schema": {"type": "string"}},
          {"name": "days", "in": "query", "description": "Use 0 for all time.", "schema": {"type": "integer", "enum": [0, 7, 30, 90], "default": 30}},
          {"name": "cursor", "in": "query", "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 25}}
        ],
        "responses": {"200": {"description": "Personalized or cold-start feed"}}
      }
    }
  },
  "components": {
    "parameters": {
      "ArxivId": {"name": "arxiv_id", "in": "path", "required": true, "schema": {"type": "string"}}
    },
    "responses": {
      "BadRequest": {"description": "Invalid query parameter"},
      "NotFound": {"description": "Paper not found"}
    },
    "schemas": {
      "Paper": {
        "type": "object",
        "required": ["arxiv_id", "version", "title", "abstract", "authors", "categories", "primary_category", "published_at", "updated_at", "links", "state"],
        "properties": {
          "arxiv_id": {"type": "string"},
          "version": {"type": "integer"},
          "title": {"type": "string"},
          "abstract": {"type": "string"},
          "authors": {"type": "array", "items": {"type": "string"}},
          "categories": {"type": "array", "items": {"type": "string"}},
          "primary_category": {"type": "string"},
          "published_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"},
          "doi": {"type": ["string", "null"]},
          "journal_ref": {"type": ["string", "null"]},
          "links": {"type": "object", "properties": {"abstract": {"type": "string", "format": "uri"}, "pdf": {"type": "string", "format": "uri"}}},
          "state": {"type": "object", "properties": {"saved": {"type": "boolean"}, "read": {"type": "boolean"}, "dismissed": {"type": "boolean"}, "last_opened_at": {"type": ["string", "null"]}}},
          "score": {"type": "number"}
        }
      }
    }
  }
}